[pyro] Adding colorized source, moving to clj-stacktrace
This commit removes pyro's own stacktrace engine in favor of using the
one built into clj-stacktrace. There's no sense in re-inventing that
particular part of the wheel, and they did a pretty bangup job of it.

We also add in an engine for printing out syntax-highlighted source code
for the relevant areas of the stacktrace, and built in the ability to
filter out stacktrace elements from the Clojure compiler, as well as
`clojure.test`, `clojure.core` and `clojure.main`. If you're a core
Clojure contributor, you can always turn those off yourself :)