[memoized] source-fn is now properly memoized
This is important because generating highlighted strings for source code
means often returning to the same input files over and over again. As
Glow's highlighting is not quick enough to avoid a noticeable hiccup, it
would be a real pain to have to re-parse each of these files each time
they're used.

The memoization we're using here comes from `core.memoize`, which is in
turn based on `core.cache`. We use a least-used caching strategy with 64
buckets, though that number might increase and/or we might make that
configurable at the top level.