[tests] Remove glow.core link in tests
We had a commented-out test that relied on glow.core, which is unusable
due to our now excluding hiccup and garden from Pyro's dependencies.

In the future I'll probably want to refactor Glow so that glow.core
doesn't immediately bring in hiccup and garden, but I don't feel like
doing that right now.