[pyro] Reload clojure.test after altering var roots
This commit reloads clojure.test after altering the var roots - this is
necessary because Clojure 1.8.0+ links clojure.test directly into
clojure.stacktrace, so we need to re-evaluate the namespace after having made our modifications.

It also makes a minor tweak in how source is printed to keep colorized
source from overflowing into the REPL session, but this doesn't fix a
number of cases and an ultimate fix will probably require a material
change in Glow upstream.